+++ base_pack/nfc_magic/scenes/nfc_magic_scene_iso15693_write.c
@@ -22,6 +22,7 @@ static void
         break;
     case Iso15693PollerEventPartial:
     case Iso15693PollerEventCardDetected:
+    case Iso15693PollerEventWriteProgress:
         break; // not emitted in Write UID mode
     }
 }